refactor: abi types belong to abi package
(Someone ought to tell me if this is non-idiomatic golang project structuring)
